backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.5.0
python-dotenv==1.0.0
pathlib==1.0.1
markdown==3.5.1
beautifulsoup4==4.12.2
requests==2.31.0
asyncio==3.4.3
pathlib2==2.3.7